Option handler for playlists and settings. Exe function to execute command files. Command aliases updated.
