+++ commands/forcestop.js
@@ -1,9 +1,10 @@
 module.exports = {
     name: 'forcestop',
     description: 'force stop command.',
+    alias: 'fs',
     cooldown: 5,
     execute(message, args, client, Discord, prefix) {
-        if (message.author.id !== '360363051792203779') return message.channel.send(':x: You are not allowed to do that!')
+        if (message.author.id !== client.global.devId) return message.channel.send(':x: You are not allowed to do that!')
         client.queue.delete(message.guild.id);
         message.channel.send('queue deleted')
     }
