+++ commands/loop.js
@@ -1,13 +1,14 @@
 module.exports = {
     name: 'loop',
     description: 'loop command.',
+    alias: 'loop',
     cooldown: 10,
     async execute(message, args, client, Discord, prefix) {
         const serverQueue = client.queue.get(message.guild.id);
         const permissions = message.channel.permissionsFor(message.author);
         const { voiceChannel } = message.member;
         if (serverQueue) {
-            if (message.author.id !== '360363051792203779') {
+            if (message.author.id !== client.global.devId) {
                 if (voiceChannel !== serverQueue.voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in the same voice channel as Musix to loop the queue!');
                 if (client.global.db.guilds[message.guild.id].permissions === true) {
                     if (client.global.db.guilds[message.guild.id].dj) {
