+++ commands/pause.js
@@ -1,6 +1,7 @@
 module.exports = {
 	name: 'pause',
 	description: 'Pause command.',
+	alias: 'pause',
 	cooldown: 5,
 	execute(message, args, client, Discord, prefix) {
 		const serverQueue = client.queue.get(message.guild.id);
@@ -8,7 +9,7 @@ module.exports = {
 		const { voiceChannel } = message.member;
 		if (serverQueue && serverQueue.playing === true) {
 			if (voiceChannel !== serverQueue.voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in the same voice channel as Musix to pause the music!');
-			if (message.author.id !== '360363051792203779') {
+			if (message.author.id !== client.global.devId) {
 				if (client.global.db.guilds[message.guild.id].permissions === true) {
 					if (client.global.db.guilds[message.guild.id].dj) {
 						if (!message.member.roles.has(client.global.db.guilds[message.guild.id].djrole)) return message.channel.send(':x: You need the `DJ` role to pause the music!');
