+++ commands/seek.js
@@ -1,6 +1,7 @@
 module.exports = {
     name: 'seek',
     description: 'seek command.',
+    alias: 'seek',
     cooldown: 10,
     async execute(message, args, client, Discord, prefix) {
         const ytdl = require('ytdl-core');
@@ -9,7 +10,7 @@ module.exports = {
         const { voiceChannel } = message.member;
         let data = await Promise.resolve(ytdl.getInfo(serverQueue.songs[0].url));
         if (serverQueue) {
-            if (message.author.id !== '360363051792203779') {
+            if (message.author.id !== client.global.devId) {
                 return message.channel.send(':x: This command is currently disabled!');
                 if (voiceChannel !== serverQueue.voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in the same voice channel as Musix to seek the song!');
                 if (client.global.db.guilds[message.guild.id].permissions === true) {
