+++ commands/skip.js
@@ -1,13 +1,14 @@
 module.exports = {
 	name: 'skip',
 	description: 'Skip command.',
+	alias: 'skip',
 	cooldown: 5,
 	execute(message, args, client, Discord, prefix) {
 		const { voiceChannel } = message.member;
 		const serverQueue = client.queue.get(message.guild.id);
 		const permissions = message.channel.permissionsFor(message.author);
 		if (!serverQueue) return message.channel.send(':x: There is nothing playing that I could skip for you.');
-		if (message.author.id !== '360363051792203779') {
+		if (message.author.id !== client.config.devId) {
 			if (voiceChannel !== serverQueue.voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in the same voice channel as Musix to skip the song!');
 			if (client.global.db.guilds[message.guild.id].permissions === true) {
 				if (client.global.db.guilds[message.guild.id].dj) {
