+++ commands/stop.js
@@ -1,12 +1,13 @@
 module.exports = {
 	name: 'stop',
 	description: 'Stop command.',
+	alias: 'stop',
 	cooldown: 5,
 	execute(message, args, client, Discord, prefix) {
 		const { voiceChannel } = message.member;
 		const serverQueue = client.queue.get(message.guild.id);
-		const permissions = message.channel.permissionsFor(message.author);	
-		if (message.author.id !== '360363051792203779') {
+		const permissions = message.channel.permissionsFor(message.author);
+		if (message.author.id !== client.config.devId) {
 			if (!serverQueue) return message.channel.send(':x: There is nothing playing that I could stop for you.');
 			if (voiceChannel !== serverQueue.voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in the same voice channel as Musix to stop the music!');
 			if (client.global.db.guilds[message.guild.id].permissions === true) {
