+++ commands/volume.js
@@ -1,6 +1,7 @@
 module.exports = {
 	name: 'volume',
 	description: 'Volume command.',
+	alias: 'volume',
 	cooldown: 5,
 	execute(message, args, client, Discord, prefix) {
 		const { voiceChannel } = message.member;
@@ -10,7 +11,7 @@ module.exports = {
 		if (!serverQueue) return message.channel.send(':x: There is nothing playing.');
 		if (!args[1]) return message.channel.send(`:loud_sound: The current volume is: **${serverQueue.volume}**`);
 		const volume = parseFloat(args[1]);
-		if (message.author.id !== '360363051792203779') {
+		if (message.author.id !== client.config.devId) {
 			if (!voiceChannel) return message.channel.send(':x: I\'m sorry but you need to be in a voice channel to change the volume!');
 			if (client.global.db.guilds[message.guild.id].permissions === true) {
 				if (client.global.db.guilds[message.guild.id].dj) {
