+++ events/message.js
@@ -2,66 +2,47 @@ module.exports = {
     name: 'message',
     async execute(client, message, Discord) {
         if (message.author.bot || !message.guild) return;
-        if (message.content === '->reset' && message.author.id === '360363051792203779') {
+        if (message.content === '->reset' && message.author.id === client.global.devId) {
             client.guilds.forEach(guild => {
                 client.global.db.guilds[guild.id] = {
-                    prefix: ">",
+                    prefix: client.config.prefix,
                     defaultVolume: 5,
                     permissions: false,
                     premium: false,
                     dj: false,
                     djrole: null
                 };
+                return message.channel.send(':white_check_mark: Reset all guild settings for __all__ guilds!');
             });
             return message.channel.send(':white_check_mark: Reset all guild settings!')
-        } else if (message.content === '->resetguildsettings' && message.author.id === '360363051792203779') {
+        } else if (message.content === '->resetguildsettings' && message.author.id === client.global.devId) {
             client.global.db.guilds[message.guild.id] = {
-                prefix: ">",
+                prefix: client.config.prefix,
                 defaultVolume: 5,
                 permissions: false,
                 premium: false,
                 dj: false,
                 djrole: null
             };
+            return message.channel.send(':white_check_mark: Reset all guild settings for this guild!');
         }
         let prefix = client.global.db.guilds[message.guild.id].prefix;
         const args = message.content.slice(prefix.length).split(' ');
+        let commandName;
         if (message.mentions.users.first()) {
-            if (message.mentions.users.first().id === '607266889537945605' && args[1] === 'help') return client.commands.get('help').execute(message, args, client, Discord, prefix, client);
-            if (message.mentions.users.first().id === '607266889537945605' && args[1] === 'prefix') return message.channel.send(`My prefix here is: \`${prefix}\`.`);
+            if (message.mentions.users.first().id === client.config.botId) {
+                if (!args[1]) return;
+                if (args[1] === 'prefix') return message.channel.send(`My prefix here is: \`${prefix}\`.`);
+                commandName = args[1].toLowerCase();
+                const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName)) || client.commandAliases.get(commandName);
+                client.funcs.exe(message, args, client, Discord, prefix, command);
+            }
         }
         if (!message.content.startsWith(prefix)) return;
         if (!args[0]) return;
-        let commandName = args[0].toLowerCase();
-        if (commandName === `p`) {
-            commandName = 'play';
-        }
-        if (commandName === 'q') {
-            commandName = 'queue';
-        }
-        if (commandName === 's') {
-            commandName = 'skip';
-        }
-        if (commandName === 'np') {
-            commandName = 'nowplaying';
-        }
-        if (commandName === 'rm') {
-            commandName = 'remove';
-        }
-        const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName));
-        const permissions = message.channel.permissionsFor(message.client.user);
-        if (!permissions.has('EMBED_LINKS')) return message.channel.send(':x: I cannot send embeds (Embed links), make sure I have the proper permissions!');
+        commandName = args[0].toLowerCase();
+        const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName)) || client.commandAliases.get(commandName);
         if (!command && message.content !== `${prefix}`) return;
-        try {
-            command.execute(message, args, client, Discord, prefix);
-        } catch (error) {
-            message.reply(`:x: there was an error trying to execute that command! Please contact support with \`${prefix}bug\`!`);
-            const embed = new Discord.RichEmbed()
-                .setTitle(`Musix ${error.toString()}`)
-                .setDescription(error.stack.replace(/at /g, '**at **'))
-                .setColor('#b50002');
-            client.fetchUser('360363051792203779').then(user => user.send(embed)).catch(console.error);
-            client.channels.get('634718645188034560').send(embed);
-        }
+        client.funcs.exe(message, args, client, Discord, prefix, command);
     }
 }
