+++ index.js
@@ -23,6 +23,9 @@ client.global = {
 };
 
 client.commands = new Discord.Collection();
+client.commandAliases = new Discord.Collection();
+client.playlistCmd = new Discord.Collection();
+client.settingCmd = new Discord.Collection();
 client.events = new Discord.Collection();
 client.cooldowns = new Discord.Collection();
 client.queue = new Map();
@@ -32,22 +35,38 @@ client.funcs.handleVideo = require('./funcs/handleVideo.js');
 client.funcs.play = require('./funcs/play.js');
 client.funcs.msToTime = require('./funcs/msToTime.js');
 client.funcs.dbget = require('./funcs/dbget.js');
+client.funcs.exe = require('./funcs/exe.js');
 
 client.config = {
   token: process.env.TOKEN,
   apikey: process.env.API_KEY,
+  botId: "607266889537945605",
+  devId: "360363051792203779",
   prefix: '>',
-  test: 'success',
 };
+
 const commandFiles = fs.readdirSync('./commands/').filter(f => f.endsWith('.js'));
 for (const file of commandFiles) {
   const command = require(`./commands/${file}`);
   client.commands.set(command.name, command);
+  client.commandAliases.set(command.alias, command);
 }
 const eventFiles = fs.readdirSync('./events/').filter(f => f.endsWith('.js'));
 for (const file of eventFiles) {
   const event = require(`./events/${file}`);
   client.events.set(event.name, event);
+}
+
+const playlistFiles = fs.readdirSync('./commands/playlist/').filter(f => f.endsWith('.js'));
+for (const file of playlistFiles) {
+  const option = require(`./commands/playlist/${file}`);
+  client.playlistCmd.set(option.name, option);
+}
+
+const settingFiles = fs.readdirSync('./commands/settings/').filter(f => f.endsWith('.js'));
+for (const file of settingFiles) {
+  const option = require(`./commands/settings/${file}`);
+  client.settingCmd.set(option.name, option);
 }
 
 client.on('ready', async () => {
